requirements.txt
absl-py==1.0.0
altair==4.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.11.4
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
cycler==0.11.0
Cython==0.29.28
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.33.3
gitdb==4.0.9
GitPython==3.1.27
glfw==2.5.3
gym==0.23.1
gym-notices==0.0.6
idna==3.3
imageio==2.19.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.4.0
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
lockfile==0.12.2
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
mypy==0.950
mypy-extensions==0.4.3
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.3
opencv-python==4.5.5.64
ortools==9.3.10497
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
PuLP==2.6.0
pure-eval==0.2.2
pyarrow==7.0.0
pycparser==2.21
pydantic==1.9.0
pydeck==0.7.1
pygame==2.1.0
Pygments==2.12.0
pylint==2.13.8
Pympler==1.0.1
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
requests==2.27.1
scipy==1.8.0
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.2.0
streamlit==1.8.1
terminado==0.13.3
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.11.0
torchinfo==1.6.5
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
wrapt==1.14.1
zipp==3.8.0